Fix bug hardcoding deploy method to config vs passed arg
Also unDRY because of subtle interdependency

+++ lib/kubert/file_access.rb
@@ -44,7 +44,7 @@ module Kubert
     end
 
     def clean_local
-      File.delete(local_path)
+      File.delete(local_path) unless local?
     end
 
     private
